requirements.txt
dbt-core==1.10.13
dbt-metricflow==0.11.0
dbt-postgres==1.9.1
dbt-semantic-interfaces==0.9.4.dev0
fastapi==0.120.1
metricflow==0.209.0
pydantic==2.12.3
psycopg2-binary==2.9.11
requests==2.32.5
uvicorn==0.38.0
